+++ chevah/compat/tests/normal/test_capabilities.py
@@ -284,7 +284,7 @@ class TestNTProcessCapabilitiesNormalUser(CompatTestCase):
         self.assertNotContains('SeBackupPrivilege:0', text)
         self.assertNotContains('SeRestorePrivilege', text)
 
-        if self.os_version in 'nt-5.1':
+        if self.os_version in ['nt-5.1']:
             # Windows XP has SE_CREATE_GLOBAL enabled even when
             # running without administrator privileges.
             self.assertContains('SeCreateGlobalPrivilege:3', text)