+++ chevah/compat/tests/normal/test_filesystem.py
@@ -2631,10 +2631,9 @@ class TestLocalFilesystemLocked(CompatTestCase, FilesystemTestMixin):
         """
         path, segments = self.tempPath()
         name = segments[-1]
-        # We use a target which we know that exists on all system and which
-        # is outside of the locked filesystem.
+        # Targets present on all systems, but outside of the locked filesystem.
         if self.os_family == 'nt':
-            target = ['c', 'windows', 'explorer.exe']
+            target = ['c', 'windows', 'system32', 'xcopy.exe']
         else:
             target = ['bin', 'cp']
         mk.fs.makeLink(